QueryBuilderHints: Use useMemo instead of useState  (#176)
I believe the source of a flaky test in Grafana is this set state being
triggered from a place where we can't wrap it with act() correctly.


https://github.com/grafana/grafana/actions/runs/16802657207/job/47587503434#step:6:91

Looking at the code, I don't see the reason why this should be state at
all - it looks like it should be just useMemo.